chore: refactoring cluster graph data to make it easier to share with Queue
